Extract /resources/... to separate modules
