+++ app/api/routes/__init__.py
@@ -1,348 +1,3 @@
-from datetime import datetime
-from os import environ
-
-from dateutil import parser
-
-import app.utils as utils
-from algoliasearch.exceptions import AlgoliaException, AlgoliaUnreachableHostException
-from app import Config, db, index
-from app.api import bp
-from app.api.auth import authenticate
-from app.api.routes import api_key, categories, languages, search
-from app.api.routes.logger import logger
-from app.api.routes.metrics import latency_summary, failures_counter
-from app.api.validations import (
-    requires_body, validate_resource, validate_resource_list, wrong_type)
-from app.models import Category, Language, Resource
-from flask import redirect, request
-from sqlalchemy import func, or_
-from sqlalchemy.exc import IntegrityError
-
-
-# Routes
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources', methods=['GET'], endpoint='get_resources')
-def resources():
-    return get_resources()
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources', methods=['POST'], endpoint='create_resources')
-@requires_body
-@authenticate
-def post_resources():
-    json = request.get_json()
-
-    if not isinstance(json, list):
-        return wrong_type("list of resources objects", type(json))
-
-    validation_errors = validate_resource_list(request.method, json)
-
-    if validation_errors:
-        return utils.standardize_response(payload=validation_errors, status_code=422)
-
-    return create_resources(json, db)
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources/<int:id>', methods=['GET'], endpoint='get_resource')
-def resource(id):
-    return get_resource(id)
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources/<int:id>', methods=['PUT'], endpoint='update_resource')
-@requires_body
-@authenticate
-def put_resource(id):
-    json = request.get_json()
-
-    if not isinstance(json, dict):
-        return wrong_type("resource object", type(json))
-
-    validation_errors = validate_resource(request.method, json, id)
-
-    if validation_errors:
-        errors = {"errors": validation_errors}
-        return utils.standardize_response(payload=errors, status_code=422)
-    return update_resource(id, request.get_json(), db)
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources/<int:id>/upvote', methods=['PUT'])
-def upvote(id):
-    return update_votes(id, 'upvotes')
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources/<int:id>/downvote', methods=['PUT'])
-def downvote(id):
-    return update_votes(id, 'downvotes')
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/resources/<int:id>/click', methods=['PUT'])
-def update_resource_click(id):
-    return add_click(id)
-
-
-# Helpers
-def get_resource(id):
-    resource = Resource.query.get(id)
-
-    if resource:
-        return utils.standardize_response(payload=dict(data=(resource.serialize)))
-
-    return redirect('/404')
-
-
-def get_resources():
-    """
-    Gets a paginated list of resources.
-
-    If the URL parameters `languages` or `category` are found
-    in the request, the list will be filtered by these parameters.
-
-    The filters are case insensitive.
-    """
-    resource_paginator = utils.Paginator(Config.RESOURCE_PAGINATOR, request)
-
-    # Fetch the filter params from the url, if they were provided.
-    languages = request.args.getlist('languages')
-    category = request.args.get('category')
-    updated_after = request.args.get('updated_after')
-    paid = request.args.get('paid')
-
-    q = Resource.query
-
-    # Filter on languages
-    if languages:
-        # Take the list of languages they pass in, join them all with OR
-        q = q.filter(
-            or_(*map(Resource.languages.any,
-                     map(Language.name.ilike, languages))
-                )
-        )
-
-    # Filter on category
-    if category:
-        q = q.filter(
-            Resource.category.has(
-                func.lower(Category.name) == category.lower()
-            )
-        )
-
-    # Filter on updated_after
-    if updated_after:
-        try:
-            uaDate = parser.parse(updated_after)
-            if uaDate > datetime.now():
-                raise Exception("updated_after greater than today's date")
-            uaDate = uaDate.strftime("%Y-%m-%d")
-        except Exception as e:
-            logger.exception(e)
-            message = 'The value for "updated_after" is invalid'
-            res = {"errors": {"unprocessable-entity": {"message": message}}}
-            return utils.standardize_response(payload=res, status_code=422)
-
-        q = q.filter(
-            or_(
-                Resource.created_at >= uaDate,
-                Resource.last_updated >= uaDate
-            )
-        )
-
-    # Filter on paid
-    if isinstance(paid, str) and paid.lower() in ['true', 'false']:
-        paidAsBool = paid.lower() == 'true'
-        q = q.filter(Resource.paid == paidAsBool)
-
-    try:
-        paginated_resources = resource_paginator.paginated_data(q)
-        if not paginated_resources:
-            return redirect('/404')
-        resource_list = [
-            resource.serialize for resource in paginated_resources.items
-        ]
-        pagination_details = resource_paginator.pagination_details(paginated_resources)
-    except Exception as e:
-        logger.exception(e)
-        return utils.standardize_response(status_code=500)
-
-    return utils.standardize_response(payload=dict(
-        data=resource_list,
-        **pagination_details))
-
-
-def get_attributes(json):
-    languages_list = Language.query.all()
-    categories_list = Category.query.all()
-
-    language_dict = {l.key(): l for l in languages_list}
-    category_dict = {c.key(): c for c in categories_list}
-
-    langs = []
-    for lang in json.get('languages') or []:
-        language = language_dict.get(lang)
-        if not language:
-            language = Language(name=lang)
-        langs.append(language)
-    categ = category_dict.get(json.get('category'), Category(name=json.get('category')))
-    return (langs, categ)
-
-
-def update_votes(id, vote_direction):
-    resource = Resource.query.get(id)
-
-    if not resource:
-        return redirect('/404')
-
-    initial_count = getattr(resource, vote_direction)
-    setattr(resource, vote_direction, initial_count + 1)
-    db.session.commit()
-
-    return utils.standardize_response(payload=dict(data=resource.serialize))
-
-
-def add_click(id):
-    resource = Resource.query.get(id)
-
-    if not resource:
-        return redirect('/404')
-
-    initial_count = getattr(resource, 'times_clicked')
-    setattr(resource, 'times_clicked', initial_count + 1)
-    db.session.commit()
-
-    return utils.standardize_response(payload=dict(data=resource.serialize))
-
-
-def update_resource(id, json, db):
-    resource = Resource.query.get(id)
-
-    if not resource:
-        return redirect('/404')
-
-    langs, categ = get_attributes(json)
-    index_object = {'objectID': id}
-
-    try:
-        logger.info(f"Updating resource. Old data: {resource.serialize}")
-        if json.get('languages'):
-            resource.languages = langs
-            index_object['languages'] = resource.serialize['languages']
-        if json.get('category'):
-            resource.category = categ
-            index_object['category'] = categ.name
-        if json.get('name'):
-            resource.name = json.get('name')
-            index_object['name'] = json.get('name')
-        if json.get('url'):
-            resource.url = json.get('url')
-            index_object['url'] = json.get('url')
-        if 'paid' in json:
-            paid = json.get('paid')
-
-            # Converts "false" and "true" to their bool
-            if type(paid) is str and paid.lower() in ["true", "false"]:
-                paid = paid.lower().strip() == "true"
-            resource.paid = paid
-            index_object['paid'] = paid
-        if 'notes' in json:
-            resource.notes = json.get('notes')
-            index_object['notes'] = json.get('notes')
-
-        try:
-            index.partial_update_object(index_object)
-
-        except (AlgoliaUnreachableHostException, AlgoliaException) as e:
-            if (environ.get("FLASK_ENV") != 'development'):
-                logger.exception(e)
-                msg = f"Algolia failed to update index for resource '{resource.name}'"
-                logger.warn(msg)
-                error = {'errors': [{"algolia-failed": {"message": msg}}]}
-                return utils.standardize_response(payload=error, status_code=500)
-
-        # Wait to commit the changes until we know that Aloglia was updated
-        db.session.commit()
-
-        return utils.standardize_response(
-            payload=dict(data=resource.serialize)
-        )
-
-    except IntegrityError as e:
-        logger.exception(e)
-        return utils.standardize_response(status_code=422)
-
-    except Exception as e:
-        logger.exception(e)
-        return utils.standardize_response(status_code=500)
-
-
-def create_resources(json, db):
-    try:
-        # Resources to return in the response
-        created_resources = []
-        # Serialized Resources to send to Algolia
-        created_resources_algolia = []
-        # Resource IDs to delete if Algolia fails to index
-        resource_id_cache = []
-
-        # Create each Resource in the database one by one
-        for resource in json:
-            langs, categ = get_attributes(resource)
-            new_resource = Resource(
-                name=resource.get('name'),
-                url=resource.get('url'),
-                category=categ,
-                languages=langs,
-                paid=resource.get('paid'),
-                notes=resource.get('notes'))
-
-            try:
-                db.session.add(new_resource)
-                db.session.commit()
-                created_resources_algolia.append(new_resource.serialize_algolia_search)
-                resource_id_cache.append(new_resource.id)
-
-            except IntegrityError as e:
-                logger.exception(e)
-                return utils.standardize_response(status_code=422)
-
-            except Exception as e:
-                logger.exception(e)
-                return utils.standardize_response(status_code=500)
-
-            created_resources.append(new_resource.serialize)
-
-        # Take all the created resources and save them in Algolia with one API call
-        try:
-            index.save_objects(created_resources_algolia)
-
-        except (AlgoliaUnreachableHostException, AlgoliaException) as e:
-            if (environ.get("FLASK_ENV") != 'development'):
-                logger.exception(e)
-
-                # Remove created resources from the db to stay in sync with Algolia
-                for res_id in resource_id_cache:
-                    res = Resource.query.get(res_id)
-                    res.languages.clear()
-                    db.session.delete(res)
-                db.session.commit()
-
-                msg = "Algolia failed to index resources"
-                error = {'errors': [{"algolia-failed": {"message": msg}}]}
-                return utils.standardize_response(payload=error, status_code=500)
-
-        # Success
-        return utils.standardize_response(payload=dict(data=created_resources))
-    except Exception as e:
-        logger.exception(e)
-        return utils.standardize_response(status_code=500)
+from app.api.routes import (
+    api_key, categories, languages, resource_creation, resource_retrieval,
+    resource_updating, search)
